Remove java7/8 directories. Currently we do not need any feature from java8.
