+++ benchmarks/build.gradle
@@ -26,7 +26,6 @@ dependencies {
     compile project(':instrumentation-java-core'),
             project(':instrumentation-java-core-impl')
             project(':instrumentation-java-core-impl-java')
-            project(':instrumentation-java-core-impl-java-8')
 }
 
 compileJmhJava {
