--- core_impl_java_7/build.gradle
@@ -1,16 +1,0 @@
-description = 'Instrumentation Core Impl Java 7'
-
-apply plugin: 'java'
-
-[compileJava, compileTestJava].each() {
-    it.sourceCompatibility = 1.7
-    it.targetCompatibility = 1.7
-}
-
-dependencies {
-    compile project(':instrumentation-java-core'),
-            project(':instrumentation-java-core-impl'),
-            project(':instrumentation-java-core-impl-java')
-
-    signature "org.codehaus.mojo.signature:java17:+@signature"
-}
