--- core_impl_java_7/src/test/java/com/google/instrumentation/stats/StatsTest.java
@@ -1,31 +1,0 @@
-/*
- * Copyright 2016, Google Inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.instrumentation.stats;
-
-import static com.google.common.truth.Truth.assertThat;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-
-/**
- * Test for accessing the {@link StatsManager} through the {@link Stats} class.
- */
-@RunWith(JUnit4.class)
-public final class StatsTest {
-  @Test
-  public void getStatsContextFactory() {
-    assertThat(Stats.getStatsContextFactory()).isNotNull();
-  }
-}
