+++ examples/build.gradle
@@ -9,7 +9,6 @@ dependencies {
     compile project(':instrumentation-java-core'),
             project(':instrumentation-java-core-impl'),
             project(':instrumentation-java-core-impl-java'),
-            project(':instrumentation-java-core-impl-java-8'),
             libraries.grpc_context
 }
